Move Navigation to VerticalNavigation
