+++ components/storybook-stories.js
@@ -37,7 +37,6 @@ export Illustration from '../components/illustration/__docs__/storybook-stories'
 export Lookup from '../components/lookup/__docs__/storybook-stories';
 export MediaObject from '../components/media-object/__docs__/storybook-stories';
 export Modal from '../components/modal/__docs__/storybook-stories';
-export Navigation from '../components/navigation/__docs__/storybook-stories';
 export Notification from '../components/notification/__docs__/storybook-stories';
 export PageHeader from '../components/page-header/__docs__/storybook-stories';
 export Panel from '../components/panel/__docs__/storybook-stories';
@@ -56,3 +55,4 @@ export TimePicker from '../components/time-picker/__docs__/storybook-stories';
 export Toast from '../components/toast/__docs__/storybook-stories';
 export Tooltip from '../components/tooltip/__docs__/storybook-stories';
 export Tree from '../components/tree/__docs__/storybook-stories';
+export VerticalNavigation from '../components/vertical-navigation/__docs__/storybook-stories';
