+++ components/vertical-navigation/__examples__/shade.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import createReactClass from 'create-react-class';
 import IconSettings from '~/components/icon-settings';
-import Navigation from '~/components/navigation';
+import VerticalNavigation from '~/components/vertical-navigation';
 
 const sampleSearchCategories = [
 	{
@@ -45,7 +45,7 @@ const Example = createReactClass({
 		return (
 			<IconSettings iconPath="/assets/icons">
 				<div style={{ width: '320px', backgroundColor: '#FAFAFB' }}>
-					<Navigation
+					<VerticalNavigation
 						id="sample-navigation"
 						variant="shade"
 						categories={sampleSearchCategories}
