+++ components/vertical-navigation/__tests__/navigation.snapshot-test.jsx
