+++ utilities/constants.js
@@ -65,8 +65,6 @@ export const MENU_DROPDOWN = 'SLDSMenuDropdown';
 export const MENU_DROPDOWN_TRIGGER = 'SLDSMenuDropdownTrigger';
 export const MENU_PICKLIST = 'SLDSMenuPicklist';
 export const MODAL = 'SLDSModal';
-export const NAVIGATION = 'SLDSNavigation';
-export const NAVIGATION_ITEM = 'SLDSNavigationItem';
 export const NOTIFICATION = 'SLDSNotification';
 export const PANEL = 'SLDSPanel';
 export const PANEL_FILTER_GROUP = 'SLDSFilterGroup';
@@ -100,3 +98,5 @@ export const TOAST_CONTAINER = 'SLDSToastContainer';
 export const TREE = 'SLDSTree';
 export const TREE_BRANCH = 'SLDSTreeBranch';
 export const TREE_ITEM = 'SLDSTreeItem';
+export const VERTICAL_NAVIGATION = 'SLDSVerticalNavigation';
+export const VERTICAL_NAVIGATION_ITEM = 'SLDSVerticalNavigationItem';
